removed old testcafe files and updated packagejson
