+++ package.json
@@ -9,6 +9,7 @@
   "dependencies": {
     "amdefine": "^1.0.1",
     "forge": "^2.3.0",
+    "selenium-webdriver": "^4.0.0-alpha.1",
     "testcafe": "^1.1.1"
   },
   "devDependencies": {
@@ -18,7 +19,7 @@
     "eslint-plugin-requirejs": "^3.1.1"
   },
   "scripts": {
-    "test": "testcafe chrome test/cafe/test.js --app \"node server/index.js\" --ssl pfx=test/cafe/certs-keys/bundle.pfx;rejectUnauthorized=true ./node_modules/.bin/eslint client"
+    "test": "node test/selenium/test.js"
   },
   "repository": {
     "type": "git",
