--- test/cafe/certs-keys/server.crt
@@ -1,11 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIBqzCCARQCCQDJKKJn2iP8gTANBgkqhkiG9w0BAQUFADAaMQswCQYDVQQGEwJV
-UzELMAkGA1UECAwCTUEwHhcNMTkwNDA4MTcwNTI0WhcNMjAwNDA3MTcwNTI0WjAa
-MQswCQYDVQQGEwJVUzELMAkGA1UECAwCTUEwgZ8wDQYJKoZIhvcNAQEBBQADgY0A
-MIGJAoGBAL3ZEEp+sytKccxNYMm+JcnDvO0wn1ARZ3GSU1HMSYLvkPK2zTmmFHpm
-6rGJ53HoUacvGKUSSvYTrsk2l83BL2w6gGg+l4a72ABDEet+H4ITbQid9tzpXkbO
-blCPtlqN2VQkS7OGAH6VIdY5HpzK5MMveYDmEHNqbFN/NA2moIwnAgMBAAEwDQYJ
-KoZIhvcNAQEFBQADgYEAKXi+JJwCrOzlVltRe6JCWr7hGRHRpUKh+PjAvq3rrWCD
-U6Pl9t23vcwaYKUQm0WfkPoU8PdEiPxAGDDINaFEFFpRbP1MU5BC4sk/ItbSsb9y
-wAkNSGf8resAcUq556dOip+dpcVoQIDdz9kE+FHY33ox4g2QN2JtPeyar0TkyOM=
------END CERTIFICATE-----
