--- test/cafe/certs-keys/server.key
@@ -1,15 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIICXgIBAAKBgQC92RBKfrMrSnHMTWDJviXJw7ztMJ9QEWdxklNRzEmC75Dyts05
-phR6Zuqxiedx6FGnLxilEkr2E67JNpfNwS9sOoBoPpeGu9gAQxHrfh+CE20Infbc
-6V5Gzm5Qj7ZajdlUJEuzhgB+lSHWOR6cyuTDL3mA5hBzamxTfzQNpqCMJwIDAQAB
-AoGBAIBnXdNk2n3/PfS2qS4KQ7ixgQjUCAy5mMOB+oSqrVW3/suKxEengtY+4aHu
-6fFq9j8jZYHQW48QRPmmyQnLDezzwZSyri5P+yLQjvqDWusgW9EppvN4KT5G3Eax
-caJH8b2flKdukGy9lB1CinV22JnTD5duKi4o0mmHBIGisTZRAkEA9UvTpTgng78F
-cAOMBoegUKyZjMVwV4dlAR2kZtt+x4yB7OHtWprjUwBCfdp32ncnMk2Kmve6jzjE
-A+GSDKrLcwJBAMYh1LdLqRvKL/LAsA4XewhCRHWadd+JkfNq/9LSoswKaw9yJH9r
-ChOWjv3taNDKSfgDC98iNwf99nu84oYEt30CQQDVyPZ2wtA9lJiKnazGbCIe7a8N
-724P+yLerBUCyD5gcGc0K1wQYXnKdyl8QpG+x6ZTwx0QHuRrus+rikcd3yVdAkBg
-w48Gwe3iQ5mtPgv9pdwRiYOzx+7/cEmWj1BHn1WgvwnQbleC3zz5mhNV0ofZSKJd
-fOdkX2pu/cpOXk5nU5BlAkEA4jhH+BXNLQ8BMtJ+9XODPLGbcLBfqFWUrJeJddPf
-ksp/IdHq/s1bz3Gj1HgKwbLd9WIUc6D4eOGvIovu9BbtCQ==
------END RSA PRIVATE KEY-----
